+++ src/git/add.ts
@@ -23,10 +23,16 @@ export class AddNote extends GitCommandsInstance {
     await this.git
       .raw(cmdList)
       .then(async () => {
-        await this.cache.loadNoteDetails(
-          parameter.repositoryPath,
-          parameter.commitHash
-        );
+        parameter.force
+          ? await this.cache.updateNoteMessage(
+              parameter.commitHash,
+              parameter.message,
+              parameter.repositoryPath
+            )
+          : await this.cache.loadNoteDetails(
+              parameter.repositoryPath,
+              parameter.commitHash
+            );
       })
       .catch((error) => {
         this.logger.error(`command error adding note: ${error}`);