+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - [#42](https://github.com/jrosco/vscode-git-notes/pull/42) ❇️ Added a new edit window class for git notes editor
 - [#43](https://github.com/jrosco/vscode-git-notes/pull/43) ❇️ Add `append` command and refactor some code to support this new command
 - [#45](https://github.com/jrosco/vscode-git-notes/pull/45) 🔊 Add `deprecated` log service for deprecated classes and methods
+- [#56](https://github.com/jrosco/vscode-git-notes/pull/56) ❇️ Create separate extension disposables classes for each extension command
 
 ### Changed